+++ compiler-core/src/type_/pattern.rs
@@ -397,7 +397,6 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
                     expected: type_.clone(),
                     situation: None,
                     location,
-                    rigid_type_names: hashmap![],
                 }),
             },
 
@@ -444,7 +443,6 @@ impl<'a, 'b> PatternTyper<'a, 'b> {
                         expected: type_,
                         situation: None,
                         location,
-                        rigid_type_names: hashmap![],
                     })
                 }
             },